Fixed issue in OptionBar.test.tsx that wouldn't allow yarn run dev
